criu: move timers dump/restore code into separate file
Fixes: #335

move timers dump/restore code into separate file
