+++ criu/include/parasite-syscall.h
@@ -22,11 +22,6 @@ struct parasite_ctl;
 struct parasite_thread_ctl;
 
 extern int parasite_dump_sigacts_seized(struct parasite_ctl *ctl, struct pstree_item *);
-extern int parasite_dump_itimers_seized(struct parasite_ctl *ctl, struct pstree_item *);
-
-struct proc_posix_timers_stat;
-extern int parasite_dump_posix_timers_seized(struct proc_posix_timers_stat *proc_args, struct parasite_ctl *ctl,
-					     struct pstree_item *);
 
 extern int parasite_dump_misc_seized(struct parasite_ctl *ctl, struct parasite_dump_misc *misc);
 extern int parasite_dump_creds(struct parasite_ctl *ctl, CredsEntry *ce);
